.ai-docs/requirements.txt
# LLM Provider SDKs (install only the one you need)
google-genai>=1.0.0,<2.0.0
openai>=1.0.0,<2.0.0
anthropic>=0.39.0,<1.0.0

# Core dependencies
pyyaml>=6.0.0,<7.0.0